requirements.txt
Django>=3.2,<5.0
numpy>=1.20
opencv-python>=4.5
mediapipe>=0.9
openpyxl>=3.0
reportlab>=3.6
google-auth>=2.20
google-api-python-client>=2.100
sounddevice>=0.4
pygetwindow>=0.0.9
pdfplumber>=0.7
python-docx>=0.8
fpdf>=1.7
google-generativeai>=0.1
djongo>=1.3.6
dnspython>=2.0
gunicorn==20.1.0
moviepy>=1.0.3
pydub>=0.25
torch>=2.0.0
openai-whisper>=20231114
transformers>=4.30.0
ffmpeg-python>=0.2.0
